requirements.txt
flask >= 0.10
whitenoise
gunicorn
kubernetes
werkzeug
Pillow
tensorflow==1.4.0
grpcio==1.7.0

